Knockoffs

Documentation for Knockoffs.

Knockoffs.Knockoff

Knockoffs.coefficient_diff

Knockoffs.coefficient_diff

Knockoffs.coefficient_diff

Knockoffs.coefficient_diff

Knockoffs.condition

Knockoffs.extract_beta

Knockoffs.fixed_knockoffs

Knockoffs.markov_knockoffs

Knockoffs.modelX_gaussian_knockoffs

Knockoffs.threshold

Knockoffs.update_normalizing_constants!

Knockoffs.Knockoff — Type

A Knockoff is an AbstractMatrix, essentially the matrix [X X̃] of concatenating X̃ to X. If A is a Knockoff, A behaves like a regular matrix and can be inputted into any function that supports inputs of type AbstractMatrix. Basic operations like @view(A[:, 1]) are supported.

Knockoffs.coefficient_diff — Function

coefficient_diff(β::AbstractVector, method=:concat)

Returns the coefficient difference statistic W[j] = |β[j]| - |β[j + p]| from a univariate (single response) regression.

Inputs

β: Vector of regression coefficients

method: Either :concatenated (default) if all knockoffs are concatenated at the end (e.g. [XX̃]) or :interleaved if each variable is immediately followed by its knockoff (e.g. [x₁x̃₁x₂x̃₂...])

Knockoffs.coefficient_diff — Function

coefficient_diff(B::AbstractMatrix)

Returns the coefficient difference statistic W[j] = |B[j, 1]| - |B[j + p, 1]| + ... + |B[j, r]| - |B[j + p, r]| from a multivariate (multiple response) regression.

Inputs

β: Matrix of regression coefficients. Each column is a vector of β.

method: Either :concatenated if all knockoffs are concatenated at the end (e.g. [XX̃]) or :interleaved if each variable is immediately followed by its knockoff (e.g. [x₁x̃₁x₂x̃₂...])

Knockoffs.coefficient_diff — Method

coefficient_diff(β::AbstractVector, original::AbstractVector{Int}, knockoff::AbstractVector{Int})

Returns the coefficient difference statistic for grouped variables W[G] = sum{j in G} |β[j]| - sum{j in G} |β[j + p]|.

Inputs

β: Vector of regression coefficients

groups: Vector storing group membership. groups[i] is the group of β[i]

original: The index of original variables in β

knockoff: The index of knockoff variables in β

Knockoffs.coefficient_diff — Method

coefficient_diff(β::AbstractVector, original::AbstractVector{Int}, knockoff::AbstractVector{Int})

Returns the coefficient difference statistic W[j] = |β[j]| - |β[j + p]| from a univariate (single response) regression, where the jth variable is stored in position original[j] of β, and its knockoff is stored in position knockoff[j]

Inputs

β: Vector of regression coefficients

original: The index of original variables in β

knockoff: The index of knockoff variables in β

Knockoffs.condition — Method

condition(x::AbstractVector, μ::AbstractVector, Σinv::AbstractMatrix, D::AbstractMatrix)

Samples a knockoff x̃ from Gaussian x using conditional distribution formulas:

If (x, x̃) ~ N((μ, μ), G) where G = [Σ Σ - D; Σ - D Σ], then we sample x̃ from x̃|x = N(x - Dinv(Σ)(x - μ), 2D - Dinv(Σ)*D)

todo: efficiency

Knockoffs.extract_beta — Method

extract_beta(β̂_knockoff::AbstractVector, fdr::Number, method::Symbol=:concatenated)

Given estimated β of original variables and their knockoffs, compute β for the original design matrix that controls the FDR.

Knockoffs.fixed_knockoffs — Method

fixed_knockoffs(X::Matrix{T}; method=:sdp)

Creates fixed knockoffs based on equation (2.2)-(2.4) of "Controlling the false discovery rate via Knockoffs" by Barber and Candes (2015)

Inputs

X: A n × p numeric matrix, each row is a sample, and each column is normalized

to mean 0 variance 1 with unit norm.

method: :equi for equi-distant knockoffs (eq 2.3) or :sdp for SDP

knockoffs (eq 2.4)

Output

Knockoff: A struct containing the original X and its knockoff X̃,

in addition to other variables (e.g. s)

Knockoffs.markov_knockoffs — Method

markov_knockoffs()

Generates knockoff of variables distributed as a discrete Markov Chain with K states.

Inputs

X: n × p matrix, each row is a sample

Q: p - 1 vector of matrices. Q[j] is a K × K matrix of transition probabilities where Q[j][l, k] = P(X{j+1} = k | X{j} = l)

q: K × 1 vector of initial probabilities

Reference

Equations 4-5 of "Gene hunting with hidden Markov model knockoffs" by Sesia, Sabatti, and Candes

Knockoffs.modelX_gaussian_knockoffs — Method

modelX_gaussian_knockoffs(X::Matrix{T})

Creates model-free multivariate normal knockoffs by sequentially sampling from conditional multivariate normal distributions.

Inputs

X: A n × p numeric matrix. Each row is a sample, and each column is standardized

to mean 0 variance 1.

Reference:

"Panning for Gold: Model-X Knockoffs for High-dimensional Controlled Variable Selection" by Candes, Fan, Janson, and Lv (2018)

Knockoffs.threshold — Method

threshold(w::AbstractVector, q::Number, method=:knockoff)

Chooses a threshold τ > 0 by setting τ = min{ t > 0 : {#j: w[j] ≤ -t} / {#j: w[j] ≥ -t} ≤ q }.

Inputs

w: Vector of feature important statistics

q: target FDR (between 0 and 1)

method: either :knockoff or :knockoff_plus

Reference:

Equation 3.10 (method=:knockoff) or 3.11 (method=:knockoff_plus) of "Panning for Gold: Model-X Knockoffs for High-dimensional Controlled Variable Selection" by Candes, Fan, Janson, and Lv (2018)

Knockoffs.update_normalizing_constants! — Method

update_normalizing_constants!(Q::AbstractMatrix{T}, q::AbstractVector{T})

Computes normalizing constants recursively using equation (5).

Inputs

Q: p - 1 vector of matrix. Q[j] is a K × K matrix of transition probabilities where Q[j][l, k] = P(X{j+1} = k | X{j} = l)

q: K × 1 vector of initial probabilities

todo: efficiency