API

Here is a list of available function calls. A detailed description can be found below.

Index

Knockoffs.coefficient_diff

Knockoffs.compare_correlation

Knockoffs.compare_pairwise_correlation

Knockoffs.extract_beta

Knockoffs.fixed_knockoffs

Knockoffs.full_knockoffscreen

Knockoffs.hmm_knockoff

Knockoffs.merge_knockoffs_with_original

Knockoffs.modelX_gaussian_knockoffs

Knockoffs.rapid

Knockoffs.threshold

Generating knockoffs

Knockoffs.fixed_knockoffs — Function

fixed_knockoffs(X::Matrix{T}; method=:sdp)

Creates fixed knockoffs based on equation (2.2)-(2.4) of "Controlling the false discovery rate via Knockoffs" by Barber and Candes (2015)

Inputs

X: A n × p numeric matrix, each row is a sample, and each column is normalized to mean 0 variance 1 with unit norm.

method: :equi for equi-distant knockoffs (eq 2.3) or :sdp for SDP knockoffs (eq 2.4)

Output

Knockoff: A struct containing the original X and its knockoff X̃, in addition to other variables (e.g. s)

Knockoffs.modelX_gaussian_knockoffs — Function

modelX_gaussian_knockoffs(X::Matrix, method::Symbol)

Creates model-free multivariate normal knockoffs by sequentially sampling from conditional multivariate normal distributions. The true mean μ and covariance Σ is estimated from data.

Inputs

X: A n × p numeric matrix. Each row is a sample, and each column is standardized

to mean 0 variance 1.

method: Either :equi, :sdp, or :asdp

Reference:

"Panning for Gold: Model-X Knockoffs for High-dimensional Controlled Variable Selection" by Candes, Fan, Janson, and Lv (2018)

todo: the SDP routine for this is quite slow for some reason. May need better

way to estimate Σ

modelX_gaussian_knockoffs(X::Matrix, method::Symbol, μ::Vector, Σ::Matrix)

Creates model-free multivariate normal knockoffs by sequentially sampling from conditional multivariate normal distributions.

Inputs

X: A n × p numeric matrix. Each row is a sample, and each column is standardized

to mean 0 variance 1.

method: Either :equi, :sdp, or :asdp

Reference:

"Panning for Gold: Model-X Knockoffs for High-dimensional Controlled Variable Selection" by Candes, Fan, Janson, and Lv (2018)

Knockoffs.hmm_knockoff — Function

hmm_knockoff(plinkname; [datadir], [plink_outfile], [fastphase_outfile], [outdir], [verbose], args...)

Generates HMM knockoffs from binary PLINK formatted files. This is done by first running fastPHASE, then running Algorithm 2 of "Gene hunting with hidden Markov model knockoffs" by Sesia, Sabatti, and Candes

Input

plinkname: Binary PLINK file names without the .bed/.bim/.fam suffix.

Optional arguments

datadir: Full path to the PLINK and fastPHASE files (default = current directory)

plink_outfile: Output PLINK format name

fastphase_outfile: The output file name from fastPHASE's alpha, theta, r files

args...: Any parameter that accepted in fastPHASE.fastphase_estim_param()

Output

plink_outfile.bed: n × p knockoff genotypes

plink_outfile.bim: SNP mapping file. Knockoff have SNP names ending in ".k"

plink_outfile.fam: Sample mapping file, this is a copy of the original plinkname.fam file

fastphase_outfile_rhat.txt: averaged r hat file from fastPHASE

fastphase_outfile_alphahat.txt: averaged alpha hat file from fastPHASE

fastphase_outfile_thetahat.txt: averaged theta hat file from fastPHASE

hmm_knockoff(snpdata::SnpData, r::AbstractVecOrMat, θ::AbstractMatrix, α::AbstractMatrix)

Generates knockoff of snpdata with loaded r, θ, α

Input

SnpData: A SnpData object from SnpArrays

r: The r vector estimated by fastPHASE

θ: The θ matrix estimated by fastPHASE

α: The α matrix estimated by fastPHASE

Optional Inputs

outdir: Output directory for generated knockoffs

plink_outfile: Output file name for knockoff genotypes

estimate_δ: If true, will estimate pseudo-FDR by computing a δ value for each SNP via likelihood ratio bound

Knockoffs.full_knockoffscreen — Function

full_knockoffscreen(x::SnpArray; windowsize::Int=100)

Generates knockoffs X̃ⱼ by on regressing Xⱼ on SNPs knockoffs within a sliding window of width windowsize.

Inputs

x: A SnpArray or String for the path of the PLINK .bed file

windowsize: Int specifying window width. Defaults to 100

Outputs

X̃: A n × p dense matrix of Float64, each row is a sample.

References

He, Zihuai, Linxi Liu, Chen Wang, Yann Le Guen, Justin Lee, Stephanie Gogarten, Fred Lu et al. "Identification of putative causal loci in whole-genome sequencing data via knockoff statistics." Nature communications 12, no. 1 (2021): 1-18.

He, Zihuai, Yann Le Guen, Linxi Liu, Justin Lee, Shiyang Ma, Andrew C. Yang, Xiaoxia Liu et al. "Genome-wide analysis of common and rare variants via multiple knockoffs at biobank scale, with an application to Alzheimer disease genetics." The American Journal of Human Genetics 108, no. 12 (2021): 2336-2353.

TODO

Use ElasticArrays.jl to avoid reallocating design matrix in each loop

Write iterator interface to avoid allocating and storing all knockoffs at once

Regular functions

Knockoffs.coefficient_diff — Function

coefficient_diff(β::AbstractVector, original::AbstractVector{Int}, knockoff::AbstractVector{Int})

Returns the coefficient difference statistic W[j] = |β[j]| - |β[j + p]| from a univariate (single response) regression, where the jth variable is stored in position original[j] of β, and its knockoff is stored in position knockoff[j]

Inputs

β: Vector of regression coefficients

original: The index of original variables in β

knockoff: The index of knockoff variables in β

coefficient_diff(β::AbstractVector, groups::Vector{Int}, original::Vector{Int}, knockoff::Vector{Int})

Returns the coefficient difference statistic for grouped variables W[G] = sum{j in G} |β[j]| - sum{j in G} |β[j + p]|.

Inputs

β: 2p × 1 vector of regression coefficients, including original and knockoff effect sizes

groups: Vector storing group membership. groups[i] is the group of β[i]

original: The index of original variables in β

knockoff: The index of knockoff variables in β

Knockoffs.threshold — Function

threshold(w::AbstractVector, q::Number, method=:knockoff)

Chooses a threshold τ > 0 by setting τ = min{ t > 0 : {#j: w[j] ≤ -t} / {#j: w[j] ≥ t} ≤ q }.

Inputs

w: Vector of feature important statistics

q: target FDR (between 0 and 1)

method: either :knockoff or :knockoff_plus

Reference:

Equation 3.10 (method=:knockoff) or 3.11 (method=:knockoff_plus) of "Panning for Gold: Model-X Knockoffs for High-dimensional Controlled Variable Selection" by Candes, Fan, Janson, and Lv (2018)

Knockoffs.extract_beta — Function

extract_beta(β̂_knockoff::Vector, fdr::Number, original::Vector{Int}, knockoff::Vector{Int}, method=:knockoff)
extract_beta(β̂_knockoff::Vector, fdr::Number, groups::Vector{Int}, original::Vector{Int}, knockoff::Vector{Int}, method=:knockoff)

Given estimated β of original variables and their knockoffs, compute β for the original design matrix that controls the FDR.

Knockoffs.compare_correlation — Function

compare_correlation()

Computes correlation between X[:, i] and X̃[:, i] for each i.

Knockoffs.compare_pairwise_correlation — Function

compare_pairwise_correlation(X::SnpArray, X̃::SnpArray, snps::Int = size(X, 2))

Computes and returns

r1: correlation between X[:, i] and X[:, j]

r2: correlation between X[:, i] and X̃[:, i]

Knockoffs.merge_knockoffs_with_original — Function

merge_knockoffs_with_original(xdata, x̃data; des::AbstractString = "knockoff")

Interleaves the original PLINK genotypes with its knockoff into a single PLINK file.

Inputs

xdata: A SnpData or Array{T, 2} of original covariates, or a String that points to the original PLINK file (without .bed/bim/fam suffix)

x̃data: A SnpData or Array{T, 2} of knockoff covariates, or a String that points to the knockoff PLINK file (without .bed/bim/fam suffix)

des: A String for output PLINK file name (without .bed/bim/fam suffix)

Outputs

xfull: A n × 2p array of original and knockoff genotypes.

original: Indices of original genotypes. original[i] is the column number for the ith SNP.

knockoff: Indices of knockoff genotypes. knockoff[i] is the column number for the ith SNP.

Missing docstring.

Missing docstring for standardize!. Check Documenter's build log for details.

Wrapper functions for SHAPEIT HMM knockoffs

There functions will eventually be replaced by Julia wrappers that no longer require user inputs.

Knockoffs.rapid — Function

rapid(rapid_exe, vcffile, mapfile, d, outfolder, w, r, s, [a])

Wrapper for the RaPID program.

Inputs

rapid_exe: Full path to the RaPID_v.1.7 executable file

vcffile: Phased VCF file name

mapfile: Map file name

d: Actual Minimum IBD length in cM

outfolder: Output folder name

w: Number of SNPs in a window for sub-sampling

r: Number of runs

s: Minimum number of successes to consider a hit

Optional Inputs

a: If true, ignore MAFs. By default (a=false) the sites are selected at random weighted by their MAFs.