API

Here is a list of available function calls. A detailed description can be found below.

Index

Knockoffs.coefficient_diff

Knockoffs.extract_beta

Knockoffs.fixed_knockoffs

Knockoffs.modelX_gaussian_knockoffs

Knockoffs.rapid

Knockoffs.threshold

Regular functions

Knockoffs.fixed_knockoffs — Function

fixed_knockoffs(X::Matrix{T}; method=:sdp)

Creates fixed knockoffs based on equation (2.2)-(2.4) of "Controlling the false discovery rate via Knockoffs" by Barber and Candes (2015)

Inputs

X: A n × p numeric matrix, each row is a sample, and each column is normalized

to mean 0 variance 1 with unit norm.

method: :equi for equi-distant knockoffs (eq 2.3) or :sdp for SDP

knockoffs (eq 2.4)

Output

Knockoff: A struct containing the original X and its knockoff X̃,

in addition to other variables (e.g. s)

Knockoffs.modelX_gaussian_knockoffs — Function

modelX_gaussian_knockoffs(X::Matrix{T})

Creates model-free multivariate normal knockoffs by sequentially sampling from conditional multivariate normal distributions.

Inputs

X: A n × p numeric matrix. Each row is a sample, and each column is standardized

to mean 0 variance 1.

Reference:

"Panning for Gold: Model-X Knockoffs for High-dimensional Controlled Variable Selection" by Candes, Fan, Janson, and Lv (2018)

Missing docstring.

Missing docstring for normalize_col!. Check Documenter's build log for details.

Knockoffs.coefficient_diff — Function

coefficient_diff(β::AbstractVector, method=:concat)

Returns the coefficient difference statistic W[j] = |β[j]| - |β[j + p]| from a univariate (single response) regression.

Inputs

β: Vector of regression coefficients

method: Either :concatenated (default) if all knockoffs are concatenated at the end (e.g. [XX̃]) or :interleaved if each variable is immediately followed by its knockoff (e.g. [x₁x̃₁x₂x̃₂...])

coefficient_diff(β::AbstractVector, original::AbstractVector{Int}, knockoff::AbstractVector{Int})

Returns the coefficient difference statistic W[j] = |β[j]| - |β[j + p]| from a univariate (single response) regression, where the jth variable is stored in position original[j] of β, and its knockoff is stored in position knockoff[j]

Inputs

β: Vector of regression coefficients

original: The index of original variables in β

knockoff: The index of knockoff variables in β

coefficient_diff(β::AbstractVector, original::AbstractVector{Int}, knockoff::AbstractVector{Int})

Returns the coefficient difference statistic for grouped variables W[G] = sum{j in G} |β[j]| - sum{j in G} |β[j + p]|.

Inputs

β: Vector of regression coefficients

groups: Vector storing group membership. groups[i] is the group of β[i]

original: The index of original variables in β

knockoff: The index of knockoff variables in β

coefficient_diff(B::AbstractMatrix)

Returns the coefficient difference statistic W[j] = |B[j, 1]| - |B[j + p, 1]| + ... + |B[j, r]| - |B[j + p, r]| from a multivariate (multiple response) regression.

Inputs

β: Matrix of regression coefficients. Each column is a vector of β.

method: Either :concatenated if all knockoffs are concatenated at the end (e.g. [XX̃]) or :interleaved if each variable is immediately followed by its knockoff (e.g. [x₁x̃₁x₂x̃₂...])

Knockoffs.threshold — Function

threshold(w::AbstractVector, q::Number, method=:knockoff)

Chooses a threshold τ > 0 by setting τ = min{ t > 0 : {#j: w[j] ≤ -t} / {#j: w[j] ≥ t} ≤ q }.

Inputs

w: Vector of feature important statistics

q: target FDR (between 0 and 1)

method: either :knockoff or :knockoff_plus

Reference:

Equation 3.10 (method=:knockoff) or 3.11 (method=:knockoff_plus) of "Panning for Gold: Model-X Knockoffs for High-dimensional Controlled Variable Selection" by Candes, Fan, Janson, and Lv (2018)

Knockoffs.extract_beta — Function

extract_beta(β̂_knockoff::AbstractVector, fdr::Number, method::Symbol=:concatenated)

Given estimated β of original variables and their knockoffs, compute β for the original design matrix that controls the FDR.

Missing docstring.

Missing docstring for extract_combine_beta. Check Documenter's build log for details.

Wrapper functions for HMM knockoffs

Missing docstring.

Missing docstring for partition. Check Documenter's build log for details.

Knockoffs.rapid — Function

rapid(rapid_exe, vcffile, mapfile, d, outfolder, w, r, s, [a])

Wrapper for the RaPID program.

Inputs

rapid_exe: Full path to the RaPID_v.1.7 executable file

vcffile: Phased VCF file name

mapfile: Map file name

d: Actual Minimum IBD length in cM

outfolder: Output folder name

w: Number of SNPs in a window for sub-sampling

r: Number of runs

s: Minimum number of successes to consider a hit

Optional Inputs

a: If true, ignore MAFs. By default (a=false) the sites are selected at random weighted by their MAFs.

Missing docstring.

Missing docstring for snpknock2. Check Documenter's build log for details.