API

Here is a list of available function calls. A detailed description can be found below.

Index

Knockoffs.approx_modelX_gaussian_knockoffs

Knockoffs.coefficient_diff

Knockoffs.compare_correlation

Knockoffs.compare_pairwise_correlation

Knockoffs.decorrelate_knockoffs

Knockoffs.extract_beta

Knockoffs.fit_lasso

Knockoffs.fixed_knockoffs

Knockoffs.full_knockoffscreen

Knockoffs.ghost_knockoffs

Knockoffs.hmm_knockoff

Knockoffs.merge_knockoffs_with_original

Knockoffs.modelX_gaussian_group_knockoffs

Knockoffs.modelX_gaussian_knockoffs

Knockoffs.normalize_col!

Knockoffs.rapid

Knockoffs.simulate_AR1

Knockoffs.solve_s

Knockoffs.solve_s_group

Knockoffs.threshold

Generating knockoffs

Knockoffs.fixed_knockoffs — Function

fixed_knockoffs(X::Matrix{T}; method=:sdp)

Creates fixed-X knockoffs.

Inputs

X: A n × p numeric matrix, each row is a sample, and each column is covariate.

method: Can be one of the following

:mvr: Minimum variance-based reconstructability knockoffs (alg 1 in ref 2)

:maxent: Maximum entropy knockoffs (alg 2 in ref 2)

:equi: Equi-distant knockoffs (eq 2.3 in ref 1),

:sdp: SDP knockoffs (eq 2.4 in ref 1)

:sdp_fast: SDP knockoffs via coordiate descent (alg 2.2 in ref 3)

kwargs...: Possible optional inputs to method, see solve_MVR, solve_max_entropy, and solve_sdp_fast

Output

Knockoff: A struct containing the original (column-normalized) X and its knockoff X̃, in addition to other variables (e.g. s)

Reference

"Controlling the false discovery rate via Knockoffs" by Barber and Candes (2015).

"Powerful knockoffs via minimizing reconstructability" by Spector, Asher, and Lucas Janson (2020)

"FANOK: Knockoffs in Linear Time" by Askari et al. (2020).

Knockoffs.modelX_gaussian_knockoffs — Function

modelX_gaussian_knockoffs(X::Matrix, method::Symbol; [covariance_approximator], [kwargs...])
modelX_gaussian_knockoffs(X::Matrix, method::Symbol, μ::Vector, Σ::Matrix; [kwargs...])

Creates model-free multivariate normal knockoffs by sequentially sampling from conditional multivariate normal distributions. The true mean μ and covariance Σ is estimated from data if not supplied.

Inputs

X: A n × p numeric matrix, each row is a sample, and each column is covariate.

method: Can be one of the following

:mvr for minimum variance-based reconstructability knockoffs (alg 1 in ref 2)

:maxent for maximum entropy knockoffs (alg 2 in ref 2)

:equi for equi-distant knockoffs (eq 2.3 in ref 1),

:sdp for SDP knockoffs (eq 2.4 in ref 1)

:sdp_fast for SDP knockoffs via coordiate descent (alg 2.2 in ref 3)

μ: A p × 1 vector of column mean of X

Σ: A p × p matrix of covariance of X

covariance_approximator: A covariance estimator, defaults to LinearShrinkage(DiagonalUnequalVariance(), :lw). See CovarianceEstimation.jl for more options.

kwargs...: Possible optional inputs to solvers specified in method, see solve_MVR, solve_max_entropy, and solve_sdp_fast

Reference:

"Panning for Gold: Model-X Knockoffs for High-dimensional Controlled Variable Selection" by Candes, Fan, Janson, and Lv (2018)

"Powerful knockoffs via minimizing reconstructability" by Spector, Asher, and Lucas Janson (2020)

"FANOK: Knockoffs in Linear Time" by Askari et al. (2020).

Covariance Approximation:

The covariance is approximated by a linear shrinkage estimator using Ledoit-Wolf with DiagonalUnequalVariance target, which seems to perform well for p>n cases. We do not simply use cov(X) since isposdef(cov(X)) is typically false. For comparison of various estimators, see: https://mateuszbaran.github.io/CovarianceEstimation.jl/dev/man/msecomp/#msecomp

Knockoffs.approx_modelX_gaussian_knockoffs — Function

approx_modelX_gaussian_knockoffs(X, method; [windowsize = 500], [covariance_approximator], kwargs...)
approx_modelX_gaussian_knockoffs(X, method, window_ranges; [covariance_approximator], kwargs...)

Generates Gaussian knockoffs by approximating the covariance as a block diagonal matrix. Each block contains windowsize consecutive features.

Inputs

X: A n × p numeric matrix or SnpArray. Each row is a sample, and each column is covariate.

method: Can be one of the following

:mvr for minimum variance-based reconstructability knockoffs (alg 1 in ref 2)

:maxent for maximum entropy knockoffs (alg 2 in ref 2)

:equi for equi-distant knockoffs (eq 2.3 in ref 1),

:sdp for SDP knockoffs (eq 2.4 in ref 1)

:sdp_fast for SDP knockoffs via coordiate descent (alg 2.2 in ref 3)

windowsize: Number of covariates to be included in a block. Each block consists of adjacent variables. The last block could contain less than windowsize variables.

window_ranges: Vector of ranges for each window. e.g. [1:97, 98:200, 201:500]

covariance_approximator: A covariance estimator, defaults to LinearShrinkage(DiagonalUnequalVariance(), :lw). See CovarianceEstimation.jl for more options.

kwargs...: Possible optional inputs to solvers specified in method, see solve_MVR, solve_max_entropy, and solve_sdp_fast

Multithreading (todo)

To enable multiple threads, simply start Julia with >1 threads and this routine will run with all available threads.

Covariance Approximation:

The covariance is approximated by a LinearShrinkageEstimator using Ledoit-Wolf shrinkage with DiagonalUnequalVariance target, which seems to perform well for p>n cases. We do not simply use cov(X) since isposdef(cov(X)) is typically false. For comparison of different estimators, see: https://mateuszbaran.github.io/CovarianceEstimation.jl/dev/man/msecomp/#msecomp

Knockoffs.modelX_gaussian_group_knockoffs — Function

modelX_gaussian_group_knockoffs(X, groups, method, Σ, μ)
modelX_gaussian_group_knockoffs(X, groups, method; [covariance_approximator])

Constructs Gaussian model-X group knockoffs. If the covariance Σ and mean μ are not specified, they will be estimated from data, i.e. we will make second-order group knockoffs. To incorporate group structure, the (true or estimated) covariance matrix is block-diagonalized according to groups membership to solve a relaxed optimization problem. See reference paper and Knockoffs.jl docs for more details.

Inputs

X: A n × p design matrix. Each row is a sample, each column is a feature.

groups: Vector of group membership

method: Method for constructing knockoffs. Options are :equi or :sdp

Σ: A p × p covariance matrix for columns of X

μ: A length p vector storing the true column means of X

covariance_approximator: A covariance estimator, defaults to LinearShrinkage(DiagonalUnequalVariance(), :lw). See CovarianceEstimation.jl for more options.

Reference

Dai & Barber 2016, The knockoff filter for FDR control in group-sparse and multitask regression

modelX_gaussian_group_knockoffs(xdata::SnpData, method)

Generates (model-X Gaussian second-order) group knockoffs for a single chromosome stored in PLINK formatted data.

todo

Handle PLINK files with multiple chromosomes and multiple plink files each storing a chromosome

Knockoffs.hmm_knockoff — Function

hmm_knockoff(plinkname; [datadir], [plink_outfile], [fastphase_outfile], [outdir], [verbose], args...)

Generates HMM knockoffs from binary PLINK formatted files. This is done by first running fastPHASE, then running Algorithm 2 of "Gene hunting with hidden Markov model knockoffs" by Sesia, Sabatti, and Candes

Input

plinkname: Binary PLINK file names without the .bed/.bim/.fam suffix.

Optional arguments

datadir: Full path to the PLINK and fastPHASE files (default = current directory)

plink_outfile: Output PLINK format name

fastphase_outfile: The output file name from fastPHASE's alpha, theta, r files

args...: Any parameter that accepted in fastPHASE.fastphase_estim_param()

Output

plink_outfile.bed: n × p knockoff genotypes

plink_outfile.bim: SNP mapping file. Knockoff have SNP names ending in ".k"

plink_outfile.fam: Sample mapping file, this is a copy of the original plinkname.fam file

fastphase_outfile_rhat.txt: averaged r hat file from fastPHASE

fastphase_outfile_alphahat.txt: averaged alpha hat file from fastPHASE

fastphase_outfile_thetahat.txt: averaged theta hat file from fastPHASE

hmm_knockoff(snpdata::SnpData, r::AbstractVecOrMat, θ::AbstractMatrix, α::AbstractMatrix)

Generates knockoff of snpdata with loaded r, θ, α

Input

SnpData: A SnpData object from SnpArrays

r: The r vector estimated by fastPHASE

θ: The θ matrix estimated by fastPHASE

α: The α matrix estimated by fastPHASE

Optional Inputs

outdir: Output directory for generated knockoffs

plink_outfile: Output file name for knockoff genotypes

estimate_δ: If true, will estimate pseudo-FDR by computing a δ value for each SNP via likelihood ratio bound

Knockoffs.full_knockoffscreen — Function

full_knockoffscreen(x::SnpArray; windowsize::Int=100)

Generates knockoffs X̃ⱼ by on regressing Xⱼ on SNPs knockoffs within a sliding window of width windowsize.

Inputs

x: A SnpArray or String for the path of the PLINK .bed file

windowsize: Int specifying window width. Defaults to 100

Outputs

X̃: A n × p dense matrix of Float64, each row is a sample.

References

He, Zihuai, Linxi Liu, Chen Wang, Yann Le Guen, Justin Lee, Stephanie Gogarten, Fred Lu et al. "Identification of putative causal loci in whole-genome sequencing data via knockoff statistics." Nature communications 12, no. 1 (2021): 1-18.

He, Zihuai, Yann Le Guen, Linxi Liu, Justin Lee, Shiyang Ma, Andrew C. Yang, Xiaoxia Liu et al. "Genome-wide analysis of common and rare variants via multiple knockoffs at biobank scale, with an application to Alzheimer disease genetics." The American Journal of Human Genetics 108, no. 12 (2021): 2336-2353.

TODO

Use ElasticArrays.jl to avoid reallocating design matrix in each loop

Write iterator interface to avoid allocating and storing all knockoffs at once

Knockoffs.ghost_knockoffs — Function

ghost_knockoffs(Zscores, Z_pos, H_pos, H, method; [windowsize], [covariance_approximator], [kwargs])

Generate Ghost knockoffs given a list of z-scores (GWAS summary statistic).

Inputs

Zscores: List of z-score statistics

Z_pos: A sorted list of SNP position for each SNP in Zscores

H_pos: A sorted list of SNP position in the reference panel H

H: A haplotype reference panel. Each row is a sample and each column is a variant.

method: Can be any of the method in approx_modelX_gaussian_knockoffs

windowsize: Number of covariates to be included in a block. Each block consists of adjacent variables. The last block could contain less than windowsize variables.

covariance_approximator: A covariance estimator, defaults to LinearShrinkage(DiagonalUnequalVariance(), :lw). See CovarianceEstimation.jl for more options.

kwargs...: Possible optional inputs to solvers specified in method, see solve_MVR, solve_max_entropy, and solve_sdp_fast

Reference

He, Z., Liu, L., Belloy, M. E., Le Guen, Y., Sossin, A., Liu, X., ... & Ionita-Laza, I. (2021). Summary statistics knockoff inference empowers identification of putative causal variants in genome-wide association studies.

Knockoffs.solve_s — Function

solve_s(Σ::AbstractMatrix, method::Symbol; kwargs...)

Solves the vector s for generating knockoffs. Σ can be a general covariance matrix but it must be wrapped in the Symmetric keyword.

Inputs

Σ: A covariance matrix (one must wrap Symmetric(Σ) explicitly)

method: Can be one of the following

:mvr for minimum variance-based reconstructability knockoffs (alg 1 in ref 2)

:maxent for maximum entropy knockoffs (alg 2 in ref 2)

:equi for equi-distant knockoffs (eq 2.3 in ref 1),

:sdp for SDP knockoffs (eq 2.4 in ref 1)

:sdp_fast for SDP knockoffs via coordiate descent (alg 2.2 in ref 3)

kwargs...: Possible optional inputs to method, see solve_MVR,

solve_max_entropy, and solve_sdp_fast

Reference

"Controlling the false discovery rate via Knockoffs" by Barber and Candes (2015).

"Powerful knockoffs via minimizing reconstructability" by Spector, Asher, and Lucas Janson (2020)

"FANOK: Knockoffs in Linear Time" by Askari et al. (2020).

Missing docstring.

Missing docstring for solve_MVR. Check Documenter's build log for details.

Missing docstring.

Missing docstring for solve_max_entropy. Check Documenter's build log for details.

Missing docstring.

Missing docstring for solve_sdp_fast. Check Documenter's build log for details.

Missing docstring.

Missing docstring for solve_SDP. Check Documenter's build log for details.

Knockoffs.solve_s_group — Function

solve_s_group(Σ, Sblocks, groups, [method=:equi]; kwargs...)

Solves the group knockoff problem, returns block diagonal matrix S satisfying 2Σ - S ⪰ 0 and the constant(s) γ.

Inputs

Σ: A covariance matrix that has been scaled to a correlation matrix.

Sblocks: A BlockDiagonal matrix that approximates Σ using group structure. Each block should be pi × pi where pi is number of variables in group i

groups: Vector of group membership

method: Method for constructing knockoffs. Options are :equi or :sdp

Regular functions

Knockoffs.coefficient_diff — Function

coefficient_diff(β::AbstractVector, original::AbstractVector{Int}, knockoff::AbstractVector{Int})

Returns the coefficient difference statistic W[j] = |β[j]| - |β[j + p]| from a univariate (single response) regression, where the jth variable is stored in position original[j] of β, and its knockoff is stored in position knockoff[j]

Inputs

β: Vector of regression coefficients

original: The index of original variables in β

knockoff: The index of knockoff variables in β

coefficient_diff(β::AbstractVector, groups::Vector{Int}, original::Vector{Int}, knockoff::Vector{Int})

Returns the coefficient difference statistic for grouped variables W[G] = sum{j in G} |β[j]| - sum{j in G} |β[j + p]|.

Inputs

β: 2p × 1 vector of regression coefficients, including original and knockoff effect sizes

groups: 2p × 1 vector storing group membership. groups[i] is the group of β[i]

original: The index of original variables in β

knockoff: The index of knockoff variables in β

Knockoffs.threshold — Function

threshold(w::AbstractVector, q::Number, method=:knockoff)

Chooses a threshold τ > 0 by setting τ = min{ t > 0 : {#j: w[j] ≤ -t} / {#j: w[j] ≥ t} ≤ q }.

Inputs

w: Vector of feature important statistics

q: target FDR (between 0 and 1)

method: either :knockoff or :knockoff_plus

Reference:

Equation 3.10 (method=:knockoff) or 3.11 (method=:knockoff_plus) of "Panning for Gold: Model-X Knockoffs for High-dimensional Controlled Variable Selection" by Candes, Fan, Janson, and Lv (2018)

Knockoffs.extract_beta — Function

extract_beta(β̂_knockoff::Vector, fdr::Number, original::Vector{Int}, knockoff::Vector{Int}, method=:knockoff)
extract_beta(β̂_knockoff::Vector, fdr::Number, groups::Vector{Int}, original::Vector{Int}, knockoff::Vector{Int}, method=:knockoff)

Given estimated β of original variables and their knockoffs, compute β for the original design matrix that controls the FDR.

Knockoffs.compare_correlation — Function

compare_correlation()

Computes correlation between X[:, i] and X̃[:, i] for each i.

Knockoffs.compare_pairwise_correlation — Function

compare_pairwise_correlation(X::SnpArray, X̃::SnpArray, snps::Int = size(X, 2))

Computes and returns

r1: correlation between X[:, i] and X[:, j]

r2: correlation between X[:, i] and X̃[:, j]

Knockoffs.merge_knockoffs_with_original — Function

merge_knockoffs_with_original(xdata, x̃data; des::AbstractString = "knockoff")

Interleaves the original PLINK genotypes with its knockoff into a single PLINK file.

Inputs

xdata: A SnpData or Array{T, 2} of original covariates, or a String that points to the original PLINK file (without .bed/bim/fam suffix)

x̃data: A SnpData or Array{T, 2} of knockoff covariates, or a String that points to the knockoff PLINK file (without .bed/bim/fam suffix)

des: A String for output PLINK file name (without .bed/bim/fam suffix)

Outputs

xfull: A n × 2p array of original and knockoff genotypes.

original: Indices of original genotypes. original[i] is the column number for the ith SNP.

knockoff: Indices of knockoff genotypes. knockoff[i] is the column number for the ith SNP.

Knockoffs.simulate_AR1 — Function

simulate_AR1(p::Int, a=1, b=1, tol=1e-3, max_corr=1, rho=nothing)

Generates p-dimensional correlation matrix for AR(1) Gaussian process, where successive correlations are drawn from Beta(a,b) independently. If rho is specified, then the process is stationary with correlation rho.

Source

https://github.com/amspector100/knockpy/blob/20eddb3eb60e0e82b206ec989cb936e3c3ee7939/knockpy/dgp.py#L61

Missing docstring.

Missing docstring for shift_until_PSD!. Check Documenter's build log for details.

Knockoffs.normalize_col! — Function

normalize_col!(X::AbstractVecOrMat)

Normalize each column of X so they sum to 1.

Knockoffs.decorrelate_knockoffs — Function

decorrelate_knockoffs(plinkfile, original, knockoff, α)

For each knockoff x̃j, we will randomly choose α% samples uniformly and set x̃j[i] ~ binomail(2, ρj) where ρj ∈ [0, 1] is the alternate allele frequency of SNP j.

Inputs

xdata: A SnpArrays.SnpData storing original and knockoff genotypes from binary PLINK trios

Optional inputs

α: A number between 0 and 1 specifying how many samples for each knockoff should be resampled (defualt 0.1)

original: Indices of original genotypes. @view(xdata.snparray[:, original]) would be the original genotype (default: entries in 2nd column bim file not ending with .k)

knockoff: Indices of knockoff genotypes. @view(xdata.snparray[:, knockoff]) would be the knockoff genotype (default: entries in 2nd column bim file ending with .k)

outfile: Output file name (defaults to "decorrelated_knockoffs")

outdir: Directory for storing output file (defaults to current directory)

Output

xnew: A n × 2p SnpArray where the knockoffs @view(xnew[:, knockoff]) have been decorrelated

Missing docstring.

Missing docstring for sample_DMC. Check Documenter's build log for details.

Knockoffs.fit_lasso — Function

fit_lasso(y, X, method=:mvr, ...)

Generates model-X knockoffs with method, runs Lasso, then applies the knockoff-filter.

Inputs

y: Response vector

X: Design matrix

method: Method for knockoff generation (defaults to :mvr)

d: Distribution of response. Defaults Normal(), for binary response (logistic regression) use Binomial().

fdrs: Target FDRs, defaults to [0.01, 0.05, 0.1, 0.25, 0.5]

filter_method: Choices are :knockoff (default) or :knockoff_plus

debias: Defines how the selected coefficients are debiased. Specify :ls for least squares (default) or :lasso for Lasso (only running on the support). To not debias, specify debias=nothing

kwargs: Additional arguments to input into glmnetcv and glmnet

todo

Allow knockoffs to be constructed based on true Σ and μ. Here we always use the second order construction which estimates Σ and μ from X.

Missing docstring.

Missing docstring for predict. Check Documenter's build log for details.

Wrapper functions for SHAPEIT HMM knockoffs

There functions will eventually be replaced by Julia wrappers that no longer require user inputs.

Knockoffs.rapid — Function

rapid(rapid_exe, vcffile, mapfile, d, outfolder, w, r, s, [a])

Wrapper for the RaPID program.

Inputs

rapid_exe: Full path to the RaPID_v.1.7 executable file

vcffile: Phased VCF file name

mapfile: Map file name

d: Actual Minimum IBD length in cM

outfolder: Output folder name

w: Number of SNPs in a window for sub-sampling

r: Number of runs

s: Minimum number of successes to consider a hit

Optional Inputs

a: If true, ignore MAFs. By default (a=false) the sites are selected at random weighted by their MAFs.